Refactor view bindings
