+++ lib/plugin/blade/button/view.ts
@@ -1,10 +1,6 @@
 import {ClassName} from '../../common/view/class-name';
-import {
-	bindDisabled,
-	bindViewProps,
-	View,
-	ViewProps,
-} from '../../common/view/view';
+import {bindDisabled, bindViewProps} from '../../common/view/reactive';
+import {View, ViewProps} from '../../common/view/view';
 import {Button} from './model/button';
 
 interface Config {
