+++ lib/plugin/common/view/view.ts
@@ -1,5 +1,4 @@
-import {SingleValueEvents, ValueMap, ValueMapEvents} from '../model/value-map';
-import {ClassName} from './class-name';
+import {ValueMap, ValueMapEvents} from '../model/value-map';
 
 interface ViewPropsObject extends Record<string, unknown> {
 	disabled: boolean;
@@ -29,67 +28,4 @@ export function defaultViewProps(
 		disabled: initialValue.disabled ?? false,
 		hidden: initialValue.hidden ?? false,
 	});
-}
-
-const className = ClassName('');
-
-function applyClass(elem: HTMLElement, className: string, active: boolean) {
-	if (active) {
-		elem.classList.add(className);
-	} else {
-		elem.classList.remove(className);
-	}
-}
-
-function extractValue<T>(ev: SingleValueEvents<T>['change']): T {
-	return ev.value;
-}
-
-function pipe<A, B, C>(
-	h1: (input: A) => B,
-	h2: (input: B) => C,
-): (input: A) => C {
-	return (input) => h2(h1(input));
-}
-
-function applyHiddenClass(elem: HTMLElement): (value: boolean) => void {
-	return (value) => {
-		applyClass(elem, className(undefined, 'hidden'), value);
-	};
-}
-
-function applyDisabledClass(elem: HTMLElement): (value: boolean) => void {
-	return (value) => {
-		applyClass(elem, className(undefined, 'disabled'), value);
-	};
-}
-
-function applyDisabled(
-	elem: HTMLButtonElement | HTMLInputElement | HTMLSelectElement,
-) {
-	return (value: boolean) => {
-		elem.disabled = value;
-	};
-}
-
-export function bindViewProps(viewProps: ViewProps, elem: HTMLElement) {
-	viewProps
-		.valueEmitter('disabled')
-		.on('change', pipe(extractValue, applyDisabledClass(elem)));
-	applyDisabledClass(elem)(viewProps.get('disabled'));
-
-	viewProps
-		.valueEmitter('hidden')
-		.on('change', pipe(extractValue, applyHiddenClass(elem)));
-	applyHiddenClass(elem)(viewProps.get('hidden'));
-}
-
-export function bindDisabled(
-	viewProps: ViewProps,
-	elem: HTMLButtonElement | HTMLInputElement | HTMLSelectElement,
-) {
-	viewProps
-		.valueEmitter('disabled')
-		.on('change', pipe(extractValue, applyDisabled(elem)));
-	applyDisabled(elem)(viewProps.get('disabled'));
 }
