+++ lib/plugin/input-bindings/common/view/text.ts
@@ -2,12 +2,8 @@ import {Formatter} from '../../../common/converter/formatter';
 import {Value} from '../../../common/model/value';
 import {ValueMap} from '../../../common/model/value-map';
 import {ClassName} from '../../../common/view/class-name';
-import {
-	bindDisabled,
-	bindViewProps,
-	View,
-	ViewProps,
-} from '../../../common/view/view';
+import {bindDisabled, bindViewProps} from '../../../common/view/reactive';
+import {View, ViewProps} from '../../../common/view/view';
 
 export type TextProps<T> = ValueMap<{
 	formatter: Formatter<T>;
