+++ lib/plugin/input-bindings/number/view/number-text.ts
@@ -4,12 +4,8 @@ import {Value, ValueEvents} from '../../../common/model/value';
 import {ValueMap} from '../../../common/model/value-map';
 import {constrainRange} from '../../../common/number-util';
 import {ClassName} from '../../../common/view/class-name';
-import {
-	bindDisabled,
-	bindViewProps,
-	View,
-	ViewProps,
-} from '../../../common/view/view';
+import {bindDisabled, bindViewProps} from '../../../common/view/reactive';
+import {View, ViewProps} from '../../../common/view/view';
 
 export type NumberTextProps = ValueMap<{
 	draggingScale: number;
