+++ lib/plugin/monitor-bindings/common/view/list.ts
@@ -2,7 +2,8 @@ import {ListItem} from '../../../common/constraint/list';
 import {createSvgIconElement} from '../../../common/dom-util';
 import {Value} from '../../../common/model/value';
 import {ClassName} from '../../../common/view/class-name';
-import {bindViewProps, View, ViewProps} from '../../../common/view/view';
+import {bindViewProps} from '../../../common/view/reactive';
+import {View, ViewProps} from '../../../common/view/view';
 
 interface Config<T> {
 	options: ListItem<T>[];
